Revert "stash apply"
This reverts commit 8fb8c6a8a9fb395de6789f1a952430a675ef3c51.
