+++ apps/web/components/experiment/experiment-locations-display.tsx
@@ -7,12 +7,9 @@ import { useTranslation } from "@repo/i18n";
 import { Card, CardHeader, CardTitle, CardContent } from "@repo/ui/components";
 import type { LocationPoint } from "@repo/ui/components/map";
 
-const EnhancedMap = dynamic(
-  () => import("@repo/ui/components").then((mod) => ({ default: mod.EnhancedMap })),
-  {
-    ssr: false,
-  },
-);
+const Map = dynamic(() => import("@repo/ui/components").then((mod) => ({ default: mod.Map })), {
+  ssr: false,
+});
 
 interface ExperimentLocationsDisplayProps {
   locations: LocationList;
@@ -102,7 +99,7 @@ export function ExperimentLocationsDisplay({
       <CardContent className="space-y-4">
         {/* Map Display */}
         <div className="overflow-hidden rounded-lg border">
-          <EnhancedMap
+          <Map
             locations={mapLocations}
             selectionMode={false}
             height="400px"
