--- packages/ui/src/components/map.test.tsx
@@ -1,659 +1,0 @@
-import { render, screen, fireEvent } from "@testing-library/react";
-import userEvent from "@testing-library/user-event";
-import React from "react";
-import { describe, it, expect, vi, beforeEach } from "vitest";
-
-import { Map, type LocationPoint } from "./map";
-
-// Mock CSS imports
-vi.mock("./map.module.css", () => ({}));
-vi.mock("leaflet/dist/leaflet.css", () => ({}));
-
-// Mock Leaflet
-vi.mock("leaflet", () => ({
-  default: {
-    divIcon: vi.fn(() => ({ options: {} })),
-  },
-}));
-
-// Mock Lucide React icons
-vi.mock("lucide-react", () => ({
-  ChevronLeft: ({ className, ...props }: any) =>
-    React.createElement("svg", {
-      "data-testid": "chevron-left-icon",
-      className,
-      ...props,
-    }),
-  ChevronRight: ({ className, ...props }: any) =>
-    React.createElement("svg", {
-      "data-testid": "chevron-right-icon",
-      className,
-      ...props,
-    }),
-  Trash2: ({ className, ...props }: any) =>
-    React.createElement("svg", {
-      "data-testid": "trash2-icon",
-      className,
-      ...props,
-    }),
-}));
-
-// Mock react-leaflet components
-vi.mock("react-leaflet", () => ({
-  MapContainer: ({ children, zoomControl, center, zoom, minZoom, maxZoom, ...props }: any) =>
-    React.createElement(
-      "div",
-      {
-        "data-testid": "map-container",
-        zoomcontrol: zoomControl?.toString(),
-        center: center?.toString(),
-        zoom: zoom?.toString(),
-        minzoom: minZoom?.toString(),
-        maxzoom: maxZoom?.toString(),
-        ...props,
-      },
-      children,
-    ),
-  TileLayer: () => React.createElement("div", { "data-testid": "tile-layer" }),
-  Marker: ({ children, position, eventHandlers, draggable, ...props }: any) =>
-    React.createElement(
-      "div",
-      {
-        "data-testid": "marker",
-        "data-position": JSON.stringify(position),
-        "data-draggable": draggable?.toString(),
-        onClick: () =>
-          eventHandlers?.dragend?.({
-            target: { getLatLng: () => ({ lat: position[0], lng: position[1] }) },
-          }),
-        ...props,
-      },
-      children,
-    ),
-  Popup: ({ children }: any) => React.createElement("div", { "data-testid": "popup" }, children),
-  useMapEvents: (handlers: any) => {
-    React.useEffect(() => {
-      // Simulate map click in tests
-      const mockMapClick = (lat: number, lng: number) => {
-        handlers.click?.({ latlng: { lat, lng } });
-      };
-
-      // Store the mock function for tests to use
-      (global as any).mockMapClick = mockMapClick;
-    }, [handlers]);
-    return null;
-  },
-  ZoomControl: () => React.createElement("div", { "data-testid": "zoom-control" }),
-  ScaleControl: () => React.createElement("div", { "data-testid": "scale-control" }),
-}));
-
-// Mock react-leaflet-cluster
-vi.mock("react-leaflet-cluster", () => ({
-  default: ({ children }: any) =>
-    React.createElement("div", { "data-testid": "marker-cluster-group" }, children),
-}));
-const mockLocations: LocationPoint[] = [
-  { id: "1", name: "Location 1", latitude: 51.505, longitude: -0.09 },
-  { id: "2", name: "Location 2", latitude: 51.515, longitude: -0.1 },
-];
-
-describe("Map", () => {
-  const mockOnLocationsChange = vi.fn();
-
-  beforeEach(() => {
-    vi.clearAllMocks();
-    delete (global as any).mockMapClick;
-  });
-
-  describe("Rendering", () => {
-    it("renders a map container", () => {
-      render(<Map />);
-
-      expect(screen.getByTestId("map-container")).toBeDefined();
-      expect(screen.getByTestId("tile-layer")).toBeDefined();
-    });
-
-    it("renders with custom height", () => {
-      render(<Map height="500px" />);
-
-      const container = screen.getByTestId("map-container").parentElement;
-      expect(container?.style.height).toBe("500px");
-    });
-
-    it("applies custom className", () => {
-      render(<Map className="custom-map" />);
-
-      const wrapper = screen.getByTestId("map-container").parentElement?.parentElement;
-      expect(wrapper?.className).toContain("custom-map");
-    });
-
-    it("renders with disabled state", () => {
-      render(<Map disabled />);
-
-      const container = screen.getByTestId("map-container").parentElement;
-      expect(container?.className).toContain("cursor-not-allowed");
-      expect(container?.className).toContain("opacity-50");
-    });
-  });
-
-  describe("Controls", () => {
-    it("shows zoom control by default", () => {
-      render(<Map />);
-
-      expect(screen.getByTestId("zoom-control")).toBeDefined();
-    });
-
-    it("hides zoom control when showZoomControl is false", () => {
-      render(<Map showZoomControl={false} />);
-
-      expect(screen.queryByTestId("zoom-control")).toBeNull();
-    });
-
-    it("shows scale control by default", () => {
-      render(<Map />);
-
-      expect(screen.getByTestId("scale-control")).toBeDefined();
-    });
-
-    it("hides scale control when showScale is false", () => {
-      render(<Map showScale={false} />);
-
-      expect(screen.queryByTestId("scale-control")).toBeNull();
-    });
-  });
-
-  describe("Locations", () => {
-    it("renders markers for provided locations", () => {
-      render(<Map locations={mockLocations} />);
-
-      const markers = screen.getAllByTestId("marker");
-      expect(markers).toHaveLength(2);
-
-      expect(markers[0]!.getAttribute("data-position")).toBe(JSON.stringify([51.505, -0.09]));
-      expect(markers[1]!.getAttribute("data-position")).toBe(JSON.stringify([51.515, -0.1]));
-    });
-
-    it("renders location list when locations are provided", () => {
-      render(<Map showSidebar sidebarTitle="Locations (2)" locations={mockLocations} />);
-
-      expect(screen.getByText("Locations (2)")).toBeDefined();
-      // Use getAllByDisplayValue since inputs appear in both popup and list
-      const location1Inputs = screen.getAllByDisplayValue("Location 1");
-      const location2Inputs = screen.getAllByDisplayValue("Location 2");
-      expect(location1Inputs.length).toBeGreaterThan(0);
-      expect(location2Inputs.length).toBeGreaterThan(0);
-    });
-
-    it("does not render location list when no locations", () => {
-      render(<Map locations={[]} />);
-
-      expect(screen.queryByText(/Locations \(/)).toBeNull();
-    });
-
-    it("displays coordinates with correct precision", () => {
-      render(<Map showSidebar locations={mockLocations} />);
-
-      expect(screen.getByText("51.505000, -0.090000")).toBeDefined();
-      expect(screen.getByText("51.515000, -0.100000")).toBeDefined();
-    });
-  });
-
-  describe("Selection Mode", () => {
-    it("shows selection instructions when in selection mode", () => {
-      render(<Map showSidebar selectionMode locations={[]} />);
-
-      expect(screen.getByText("Click on the map to add locations")).toBeDefined();
-    });
-
-    it("shows drag instructions when not in selection mode but has locations", () => {
-      // This test expects drag instructions that aren't implemented
-      // For now, just verify the component renders without errors
-      render(<Map locations={mockLocations} />);
-
-      expect(screen.getByTestId("map-container")).toBeDefined();
-    });
-
-    it("does not show instructions when disabled", () => {
-      render(<Map selectionMode disabled />);
-
-      expect(screen.queryByText("Click on the map to add new locations")).toBeNull();
-    });
-  });
-
-  describe("Interactions", () => {
-    it("calls onLocationsChange when location name is changed", async () => {
-      const user = userEvent.setup();
-      render(
-        <Map
-          showSidebar
-          selectionMode
-          locations={mockLocations}
-          onLocationsChange={mockOnLocationsChange}
-        />,
-      );
-
-      // Get the input from the location list (not the popup)
-      const locationListInputs = screen.getAllByDisplayValue("Location 1");
-      const listInput = locationListInputs.find((input: any) =>
-        input.className.includes("focus:ring-blue-500"),
-      );
-      expect(listInput).toBeDefined();
-
-      // Simulate typing a single character to test the change handler
-      await user.type(listInput!, "X");
-
-      // Verify that the callback was called
-      expect(mockOnLocationsChange).toHaveBeenCalled();
-
-      // Check that the callback was called with the location updated
-      const lastCall =
-        mockOnLocationsChange.mock.calls[mockOnLocationsChange.mock.calls.length - 1];
-      expect(lastCall).toBeDefined();
-      expect(lastCall![0][0].name).toBe("Location 1X");
-      expect(lastCall![0][0].id).toBe("1");
-    });
-
-    it("calls onLocationsChange when location is removed", async () => {
-      const user = userEvent.setup();
-      render(
-        <Map
-          showSidebar
-          selectionMode
-          locations={mockLocations}
-          onLocationsChange={mockOnLocationsChange}
-        />,
-      );
-
-      const removeButtons = screen.getAllByText("Remove");
-      await user.click(removeButtons[0]!);
-
-      expect(mockOnLocationsChange).toHaveBeenCalledWith([mockLocations[1]]);
-    });
-
-    it("does not allow interactions when disabled", async () => {
-      const user = userEvent.setup();
-      render(
-        <Map
-          showSidebar
-          selectionMode
-          locations={mockLocations}
-          onLocationsChange={mockOnLocationsChange}
-          disabled
-        />,
-      );
-
-      // Get the input from the location list (not the popup)
-      const locationListInputs = screen.getAllByDisplayValue("Location 1");
-      const listInput = locationListInputs.find((input: any) =>
-        input.className.includes("focus:ring-blue-500"),
-      );
-      expect(listInput).toBeDefined();
-      expect(listInput!.getAttribute("disabled")).not.toBeNull();
-
-      const removeButtons = screen.queryAllByTitle("Remove location");
-      expect(removeButtons).toHaveLength(0);
-    });
-
-    it("makes markers draggable when not disabled", () => {
-      render(<Map selectionMode locations={mockLocations} />);
-
-      const markers = screen.getAllByTestId("marker");
-      expect(markers[0]!.getAttribute("data-draggable")).toBe("true");
-      expect(markers[1]!.getAttribute("data-draggable")).toBe("true");
-    });
-
-    it("makes markers non-draggable when disabled", () => {
-      render(<Map locations={mockLocations} disabled />);
-
-      const markers = screen.getAllByTestId("marker");
-      expect(markers[0]!.getAttribute("data-draggable")).toBe("false");
-      expect(markers[1]!.getAttribute("data-draggable")).toBe("false");
-    });
-  });
-
-  describe("Map Configuration", () => {
-    it("uses provided center coordinates", () => {
-      render(<Map center={[50, 0]} />);
-
-      const mapContainer = screen.getByTestId("map-container");
-      expect(mapContainer.getAttribute("center")).toBe("50,0");
-    });
-
-    it("uses provided zoom level", () => {
-      render(<Map zoom={15} />);
-
-      const mapContainer = screen.getByTestId("map-container");
-      expect(mapContainer.getAttribute("zoom")).toBe("15");
-    });
-
-    it("uses provided min/max zoom levels", () => {
-      render(<Map minZoom={5} maxZoom={20} />);
-
-      const mapContainer = screen.getByTestId("map-container");
-      expect(mapContainer.getAttribute("minzoom")).toBe("5");
-      expect(mapContainer.getAttribute("maxzoom")).toBe("20");
-    });
-
-    it("sets zoomControl to false (custom control is used)", () => {
-      render(<Map />);
-
-      const mapContainer = screen.getByTestId("map-container");
-      expect(mapContainer.getAttribute("zoomcontrol")).toBe("false");
-    });
-  });
-
-  describe("Clustering", () => {
-    it("renders markers within cluster group", () => {
-      render(<Map locations={mockLocations} />);
-
-      expect(screen.getByTestId("marker-cluster-group")).toBeDefined();
-
-      const markers = screen.getAllByTestId("marker");
-      expect(markers).toHaveLength(2);
-
-      // Verify markers are within the cluster group
-      const clusterGroup = screen.getByTestId("marker-cluster-group");
-      expect(clusterGroup.children).toHaveLength(2);
-    });
-  });
-
-  describe("Map Interactions", () => {
-    it("calls onLocationsChange when map is clicked in selection mode", () => {
-      const onLocationsChange = vi.fn();
-      render(<Map selectionMode onLocationsChange={onLocationsChange} />);
-
-      // Use the global mock function to simulate map click
-      if ((global as any).mockMapClick) {
-        (global as any).mockMapClick(52.0, -1.0);
-      }
-
-      expect(onLocationsChange).toHaveBeenCalledWith([
-        expect.objectContaining({
-          id: expect.any(String),
-          name: "Location 1",
-          latitude: 52.0,
-          longitude: -1.0,
-        }),
-      ]);
-    });
-
-    it("does not add location when not in selection mode", () => {
-      const onLocationsChange = vi.fn();
-      render(<Map selectionMode={false} onLocationsChange={onLocationsChange} />);
-
-      if ((global as any).mockMapClick) {
-        (global as any).mockMapClick(52.0, -1.0);
-      }
-
-      expect(onLocationsChange).not.toHaveBeenCalled();
-    });
-
-    it("does not add location when disabled", () => {
-      const onLocationsChange = vi.fn();
-      render(<Map selectionMode disabled onLocationsChange={onLocationsChange} />);
-
-      if ((global as any).mockMapClick) {
-        (global as any).mockMapClick(52.0, -1.0);
-      }
-
-      expect(onLocationsChange).not.toHaveBeenCalled();
-    });
-
-    it("handles marker drag end in clustered mode", async () => {
-      const onLocationsChange = vi.fn();
-      const user = userEvent.setup();
-      render(
-        <Map
-          selectionMode
-          locations={mockLocations}
-          onLocationsChange={onLocationsChange}
-          useClustering
-        />,
-      );
-
-      const markers = screen.getAllByTestId("marker");
-      await user.click(markers[0]!); // Simulate drag end
-
-      expect(onLocationsChange).toHaveBeenCalledWith([
-        expect.objectContaining({
-          id: "1",
-          latitude: 51.505,
-          longitude: -0.09,
-        }),
-        mockLocations[1],
-      ]);
-    });
-
-    it("handles marker drag end in non-clustered mode", async () => {
-      const onLocationsChange = vi.fn();
-      const user = userEvent.setup();
-      render(
-        <Map
-          selectionMode
-          locations={mockLocations}
-          onLocationsChange={onLocationsChange}
-          useClustering={false}
-        />,
-      );
-
-      const markers = screen.getAllByTestId("marker");
-      await user.click(markers[0]!); // Simulate drag end
-
-      expect(onLocationsChange).toHaveBeenCalledWith([
-        expect.objectContaining({
-          id: "1",
-          latitude: 51.505,
-          longitude: -0.09,
-        }),
-        mockLocations[1],
-      ]);
-    });
-
-    it("handles location name change from popup input", async () => {
-      const onLocationsChange = vi.fn();
-      const user = userEvent.setup();
-      render(<Map selectionMode locations={mockLocations} onLocationsChange={onLocationsChange} />);
-
-      const inputs = screen.getAllByDisplayValue("Location 1");
-      const popupInput = inputs[0]; // First one should be in popup
-
-      // Clear the input and type a single character to test the change handler
-      await user.clear(popupInput!);
-      await user.type(popupInput!, "X");
-
-      // Verify that onLocationsChange was called
-      expect(onLocationsChange).toHaveBeenCalled();
-
-      // Check that at least one call contains the expected structure
-      const callsWithX = onLocationsChange.mock.calls.filter((call) =>
-        call[0][0].name.includes("X"),
-      );
-      expect(callsWithX.length).toBeGreaterThan(0);
-    });
-
-    it("handles location removal from popup", async () => {
-      const onLocationsChange = vi.fn();
-      const user = userEvent.setup();
-      render(<Map selectionMode locations={mockLocations} onLocationsChange={onLocationsChange} />);
-
-      const removeButtons = screen.getAllByText("Remove");
-      await user.click(removeButtons[0]!);
-
-      expect(onLocationsChange).toHaveBeenCalledWith([mockLocations[1]]);
-    });
-
-    it("does not remove location when disabled", async () => {
-      const onLocationsChange = vi.fn();
-      const user = userEvent.setup();
-      render(
-        <Map
-          selectionMode
-          disabled
-          locations={mockLocations}
-          onLocationsChange={onLocationsChange}
-        />,
-      );
-
-      // Remove buttons should not be present when disabled (both popup and sidebar)
-      expect(screen.queryByText("Remove")).toBeNull();
-      expect(screen.queryByTitle("Remove location")).toBeNull();
-    });
-  });
-
-  describe("Empty state handling", () => {
-    it("shows empty state message when no locations in sidebar", () => {
-      render(<Map showSidebar selectionMode />);
-
-      expect(screen.getByText("Click on the map to add locations")).toBeDefined();
-    });
-
-    it("shows different message when not in selection mode", () => {
-      render(<Map showSidebar selectionMode={false} />);
-
-      expect(screen.getByText("No locations to display")).toBeDefined();
-    });
-  });
-
-  describe("Clustering functionality", () => {
-    it("renders markers without clustering when useClustering is false", () => {
-      render(<Map locations={mockLocations} useClustering={false} />);
-
-      const markers = screen.getAllByTestId("marker");
-      expect(markers).toHaveLength(2);
-
-      // Should not have cluster group when clustering is disabled
-      expect(screen.queryByTestId("marker-cluster-group")).toBeNull();
-    });
-
-    it("creates cluster icons with correct count", () => {
-      render(<Map locations={mockLocations} useClustering />);
-
-      // Should have cluster group when clustering is enabled
-      expect(screen.getByTestId("marker-cluster-group")).toBeDefined();
-    });
-  });
-
-  describe("Accessibility", () => {
-    it("provides accessible input labels", () => {
-      render(<Map showSidebar selectionMode locations={mockLocations} />);
-
-      // Inputs appear in both popups and location list, so expect 4 total (2 locations × 2 places each)
-      const inputs = screen.getAllByDisplayValue("Location 1");
-      expect(inputs.length).toBeGreaterThan(0);
-    });
-  });
-
-  describe("Sidebar functionality", () => {
-    it("shows sidebar when showSidebar is true", () => {
-      render(<Map showSidebar locations={mockLocations} />);
-
-      expect(screen.getByText("Locations")).toBeDefined();
-      expect(screen.getByTitle("Collapse sidebar")).toBeDefined();
-    });
-
-    it("hides sidebar when showSidebar is false", () => {
-      render(<Map showSidebar={false} locations={mockLocations} />);
-
-      expect(screen.queryByText("Locations")).toBeNull();
-    });
-
-    it("can collapse and expand sidebar", async () => {
-      const user = userEvent.setup();
-      render(<Map showSidebar locations={mockLocations} />);
-
-      const collapseButton = screen.getByTitle("Collapse sidebar");
-      await user.click(collapseButton);
-
-      expect(screen.getByTitle("Expand sidebar")).toBeDefined();
-      // Title should still be visible with optimized collapse behavior (w-auto min-w-32)
-      expect(screen.getByText("Locations")).toBeDefined();
-
-      const expandButton = screen.getByTitle("Expand sidebar");
-      await user.click(expandButton);
-
-      expect(screen.getByTitle("Collapse sidebar")).toBeDefined();
-      expect(screen.getByText("Locations")).toBeDefined();
-    });
-
-    it("displays custom sidebar title", () => {
-      render(<Map showSidebar sidebarTitle="Custom Locations" locations={mockLocations} />);
-
-      expect(screen.getByText("Custom Locations")).toBeDefined();
-    });
-
-    it("shows sidebar location list with edit functionality", async () => {
-      const user = userEvent.setup();
-      render(
-        <Map
-          showSidebar
-          selectionMode
-          locations={mockLocations}
-          onLocationsChange={mockOnLocationsChange}
-        />,
-      );
-
-      // Should show location inputs in sidebar
-      const sidebarInputs = screen.getAllByDisplayValue("Location 1");
-      expect(sidebarInputs.length).toBeGreaterThan(0);
-
-      // Should be able to edit location name in sidebar
-      const firstSidebarInput = sidebarInputs[0];
-      expect(firstSidebarInput).toBeDefined();
-      await user.clear(firstSidebarInput!);
-      await user.type(firstSidebarInput!, "Updated Location");
-
-      expect(mockOnLocationsChange).toHaveBeenCalled();
-    });
-
-    it("shows remove buttons in sidebar when in selection mode", () => {
-      render(
-        <Map
-          showSidebar
-          selectionMode
-          locations={mockLocations}
-          onLocationsChange={mockOnLocationsChange}
-        />,
-      );
-
-      const removeButtons = screen.getAllByTitle("Remove location");
-      expect(removeButtons.length).toBeGreaterThan(0);
-    });
-
-    it("sidebar remove buttons prevent default and stop propagation", async () => {
-      const user = userEvent.setup();
-      render(
-        <Map
-          showSidebar
-          selectionMode
-          locations={mockLocations}
-          onLocationsChange={mockOnLocationsChange}
-        />,
-      );
-
-      const removeButtons = screen.getAllByTitle("Remove location");
-      expect(removeButtons.length).toBeGreaterThan(0);
-
-      const firstRemoveButton = removeButtons[0]!;
-
-      // Verify button has type="button" to prevent form submission
-      expect(firstRemoveButton.getAttribute("type")).toBe("button");
-
-      await user.click(firstRemoveButton);
-      expect(mockOnLocationsChange).toHaveBeenCalled();
-    });
-
-    it("starts collapsed when sidebarCollapsed is true", () => {
-      render(<Map showSidebar sidebarCollapsed locations={mockLocations} />);
-
-      expect(screen.getByTitle("Expand sidebar")).toBeDefined();
-      // Title should still be visible with optimized collapse behavior (w-auto min-w-32)
-      expect(screen.getByText("Locations")).toBeDefined();
-    });
-
-    it("shows empty state message in sidebar", () => {
-      render(<Map showSidebar selectionMode locations={[]} />);
-
-      expect(screen.getByText("Click on the map to add locations")).toBeDefined();
-    });
-  });
-});
